aug-0.2.2
git-svn-id: https://aug.svn.sourceforge.net/svnroot/aug/trunk@5 a3a97f7c-8e16-0410-9893-1d5f7d0f8f6e

+++ src/c/augnet/utility.c
@@ -17,7 +17,7 @@ aug_addbuf(struct aug_buf* buf, size_t num)
     if (buf->size_ && num) {
 
         buf->iov_->iov_base = (char*)buf->iov_->iov_base + num;
-        buf->iov_->iov_len -= num;
+        buf->iov_->iov_len -= (int)num;
         num = 0;
     }
 }
